// src/text-editor/enums/editor.command.ts
export enum EditorCommand {
  Bold = 'bold',
  Indent = 'indent',
  Italic = 'italic',
  JustifyCenter = 'justifyCenter',
  JustifyFull = 'justifyFull',
  JustifyLeft = 'justifyLeft',
  JustifyRight = 'justifyRight',
  Outdent = 'outdent',
  RemoveFormat = 'removeFormat',
  StrikeThrough = 'strikeThrough',
  Underline = 'underline',
  ForeColor = 'foreColor',
  StyleWithCSS = 'styleWithCSS'
}